+++ server/utils/jwt.js
@@ -21,6 +21,10 @@ function getJWTPayload(token) {
 export async function getUserForJWT(token: string): Promise<User> {
   const payload = getJWTPayload(token);
 
+  if (payload.type === "email-signin") {
+    throw new AuthenticationError("Invalid token");
+  }
+
   // check the token is within it's expiration time
   if (payload.expiresAt) {
     if (new Date(payload.expiresAt) < new Date()) {
